Updates GradientSelect component
